requirements.txt
asgiref==3.3.1
Django==3.1.7
django-jsonfield==1.4.1
django-simple-history==2.12.0
djangorestframework==3.12.2
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
